Fixing favicon and adding social links
In this commit I fixed the favicon in the tab to be an m rather than a d.
favicon's are just images, the browser grabs whatever thing is named favicon.ico
in the root director. I went back to the original repository this was forked
from, looked at all the other forks until I found one from a user that started
with m, and then copied the icon they'd created.

Then I added icons for linked in and github. I corrected the linked in one by
changing the class to fab, since that's what it said on fontawesomes website.
Then I did a search at the "Find your icons from here" link above to find the
github icon